+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - List
    - fmap (map)
    - operator[]
+   - mempty and mappend
 
 ## [0.3.0] - 2020-11-09
 ### Added